Separated each function into its respective file
